+++ python/metatensor-learn/CHANGELOG.md
@@ -23,6 +23,9 @@ follows [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   `LayerNorm`, `InvariantLayerNorm`, `EquivariantLinear`, `Sequential`, `Tanh`,
   and `InvariantTanh` (#513)
 
+### Fixed
+- Set correct device for output of when torch default device is different than input device (#595)
+
 ## [Version 0.2.1](https://github.com/lab-cosmo/metatensor/releases/tag/metatensor-learn-v0.2.1) - 2024-03-01
 
 ### Changed